Changes made to get the game loop to behave.
PlayerController.cs made Key.L take 2 away from health.
GameManager.cs moved the buttons to StartScreen.cs and in Update() if scene is Game Over or Welcome then Destroy PlayerController and set it to null.
Removed the singleton and manager from GameOver and just had it call buttons as appropriate